model_server/requirements.txt
fastapi
uvicorn
transformers
accelerate
diffusers
safetensors
pillow
torch
torchvision
torchaudio
